+++ CHANGELOG.md
@@ -2,8 +2,51 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
-<a name="1.3.0"></a>
-# [1.3.0](https://github.com/biancode/node-red-contrib-modbus/compare/v1.2.0...v1.3.0) (2017-09-29)
+<a name="2.0.1-beta.5"></a>
+## [2.0.1-beta.5](https://github.com/biancode/node-red-contrib-modbus/compare/v2.0.1-beta.4...v2.0.1-beta.5) (2017-10-02)
+
+
+
+<a name="2.0.1-beta.4"></a>
+## [2.0.1-beta.4](https://github.com/biancode/node-red-contrib-modbus/compare/v2.0.1-beta.3...v2.0.1-beta.4) (2017-10-02)
+
+* breaking change for the queue info node to work without the Node-RED catch node
+* new Modbus Flex Server ( flexible server of modbus-serial )
+* upgrade to serialport 5.x
+* upgrade to node-modbus 4.x
+* upgrade to debug 3.x
+* using vm2 sandboxes
+* locales updates
+
+<a name="2.0.1-beta.3"></a>
+## [2.0.1-beta.3](https://github.com/biancode/node-red-contrib-modbus/compare/v2.0.1-beta.2...v2.0.1-beta.3) (2017-10-02)
+
+
+
+<a name="2.0.1-beta.2"></a>
+## [2.0.1-beta.2](https://github.com/biancode/node-red-contrib-modbus/compare/v2.0.1-beta.1...v2.0.1-beta.2) (2017-09-29)
+
+
+
+<a name="2.0.1-beta.1"></a>
+## [2.0.1-beta.1](https://github.com/biancode/node-red-contrib-modbus/compare/v1.0.16...v2.0.1-beta.1) (2017-09-29)
+
+
+
+<a name="2.0.1-beta.0"></a>
+## [2.0.1-beta.0](https://github.com/biancode/node-red-contrib-modbus/compare/v2.0.1-alpha.1...v2.0.1-beta.0) (2017-09-26)
+
+
+
+<a name="2.0.1-alpha.1"></a>
+## [2.0.1-alpha.1](https://github.com/biancode/node-red-contrib-modbus/compare/v2.0.1-alpha.0...v2.0.1-alpha.1) (2017-09-26)
+
+
+
+<a name="2.0.1-alpha.0"></a>
+## [2.0.1-alpha.0](https://github.com/biancode/node-red-contrib-modbus/compare/v1.2.2...v2.0.1-alpha.0) (2017-09-23)
+
+
 
 <a name="1.2.0"></a>
 # [1.2.0](https://github.com/biancode/node-red-contrib-modbus/compare/v1.1.0...v1.2.0) (2017-07-19)